+++ internal/auth/tokens/models.go
@@ -37,6 +37,7 @@ type Token struct {
 	ID          int           `db:"id" goqu:"skipinsert,skipupdate"`
 	UID         string        `db:"uid"`
 	UserID      *int          `db:"user_id"`
+	ClientID    *int          `db:"client_id"`
 	Created     time.Time     `db:"created" goqu:"skipupdate"`
 	LastUsed    *time.Time    `db:"last_used"`
 	Expires     *time.Time    `db:"expires"`
@@ -102,8 +103,10 @@ func (m *Manager) Create(token *Token) error {
 		return errors.New("no application")
 	}
 
-	token.Created = time.Now()
-	token.UID = base58.NewUUID()
+	token.Created = time.Now().UTC()
+	if token.UID == "" {
+		token.UID = base58.NewUUID()
+	}
 
 	ds := db.Q().Insert(TableName).
 		Rows(token).